Add revcloudtext command
This command draws a rectangle around an MTEXT object and converts it
to a revcloud. This is like the first project I tried when I started
learning AutoLISP and I've finally gotten it to work. It feels very
much like a victory.

I also deleted my old attempts at the command, since they're
unnecessary now.