Factor out redundant code into a define-dimension-command macro
Okay, so that's not quite true, but for AutoLISP it might as well
be. define-dimension-command is a function that serves the same
purpose as a macro in Common Lisp: it performs transformations on Lisp
code and evaluates it. It's a bit of a pain without the ubiquitous
defmacro-backquote syntax, but this is the best it's gonna get for
AutoLISP.

This is a huge breakthrough.