Add prune command to remove redundant vertices from dense polylines
I had a project where I traced over a photograph of a mountain using
Bézier curves in Blender, then converted it to an SVG, then a DXF,
then a DWG. The mountain outline was ridiculously dense, and took
fifteen minutes to cut out. The prune command will determine which
vertices aren't necessary to have in the outline and remove them
accordingly. The algorithm may change with additional testing.