Refactor prune command
The angle-difference algorithm is, unfortunately, constrained by low
precision in AutoLISP's floating-point arithmetic. The variable
*EPSILON* serves no purpose, since we can't calculate to a usable
precision anyway.

Removed some commented-out code that never should have made it into a
commit anyway

Refactored the prune-points loop to make it more readable

I think that about sums it up. The algorithm is more or less finalized
for now, until such a time comes that it needs to be tweaked.